+++ index.php
@@ -9,11 +9,11 @@ use Rammewerk\Router\Router;
 
 require __DIR__ . '/vendor/autoload.php';
 
-$router = new Router(static fn($class) => new $class());
-$router->add('/', \Rammewerk\Router\Tests\Fixtures\ParameterTestRoute::class)->classMethod('stringTest');
-
-echo $router->dispatch('/kristoffer');
-die;
+//$router = new Router(static fn($class) => new $class());
+//$router->add('/', \Rammewerk\Router\Tests\Fixtures\ParameterTestRoute::class)->classMethod('stringTest');
+//
+//echo $router->dispatch('/kristoffer');
+//die;
 
 $memoryBefore = memory_get_usage();                                            // Memory usage before the test
 $memoryPeakBefore = memory_get_peak_usage() / 1024 / 1024;                     // Peak memory after the test