+++ tests/Benchmark/BenchmarkTest.php
@@ -9,6 +9,9 @@ use Rammewerk\Router\Foundation\RouteUtility;
 use Rammewerk\Router\Router;
 use Rammewerk\Router\Tests\Fixtures\ParameterTestRoute;
 
+class ClassRoute {}
+
+
 class BenchmarkTest extends Benchmark {
 
     protected int $iterations = 1000000;
@@ -20,17 +23,26 @@ class BenchmarkTest extends Benchmark {
         $route = new RouteDefinition('/', ParameterTestRoute::class)->classMethod('stringTest');
         $instance = new ParameterTestRoute();
 
-        $this->benchmark('call_user_func_array', function () use ($instance, $route) {
-            $result = call_user_func([$instance, $route->classMethod], 'hello');
-            if (!$result === 'hello') {
-                throw new \Exception('Invalid result');
+        $handler = ParameterTestRoute::class;
+        $handler_class = new ClassRoute();
+
+        $this->benchmark('is_array', function () use ($handler) {
+            if (is_array($handler)) {
+                throw new \Exception('Failed');
             }
         });
 
-        $this->benchmark('call', function () use ($instance, $route) {
-            $result = $instance->{$route->classMethod}(...['hello']);
-            if (!$result === 'hello') {
-                throw new \Exception('Invalid result');
+        $this->benchmark('is_string', function () use ($handler) {
+            if (!is_string($handler)) {
+                throw new \Exception('Failed');
+            }
+        });
+
+        $this->benchmark('isClass', function () use ($handler_class) {
+            if ($handler_class instanceof ClassRoute) {
+
+            } else {
+                throw new \Exception('Failed');
             }
         });